requirements.txt
requests
beautifulsoup4
html5lib
google-api-python-client
flask
virtualenv
gunicorn
pyyaml
